+++ src/transport/onion.js
@@ -14,16 +14,25 @@ module.exports = async (app, dir=defaultDir, hs_dir=path.join(dir, 'hidden_servi
   // Get the granax lib, install on demand if its the first time we're using it.
   const granax = await require('./granax-dep')()
 
+  const torConfig = {
+    DataDirectory: dir
+  , HiddenServiceDir: hs_dir
+  , HiddenServiceVersion: 3
+  , HiddenServicePort: `80 127.0.0.1:${httpPort}`
+  }
+
+  if (process.env.ONION_NONANONYMOUS) {
+    console.log('[warn] Tor Hidden Service configured to run in NON-ANONYMOUS mode. It will provide no privacy.')
+    Object.assign(torConfig, {
+      SocksPort: 0
+    , HiddenServiceNonAnonymousMode: 1
+    , HiddenServiceSingleHopMode: 1
+    })
+  }
+
   // Setup an hidden service over the HTTP server
   return new Promise((resolve, reject) =>
-    granax({ authOnConnect: true  }, [
-      { DataDirectory: dir }
-    , {
-        HiddenServiceDir: hs_dir
-      , HiddenServiceVersion: 3
-      , HiddenServicePort: `80 127.0.0.1:${httpPort}`
-      }
-    ])
+    granax({ authOnConnect: true  }, [ torConfig ])
     .on('error', reject)
     .on('ready', _ => resolve(`http://${getHost(hs_dir)}`))
   )